requirements.txt
flask==3.0.2
groq==0.4.2
firebase-admin==6.5.0
python-dotenv==1.0.1
Flask-CORS==4.0.0 